Free/free_pg_baseline/requirements.txt
psycopg2-binary>=2.9
pyahocorasick>=2.0.0
pyroaring>=0.3.3
